+++ CmriSubroutines/Subroutines.cs
@@ -8,6 +8,7 @@ using System.IO;
 using System.Net;
 using System.Net.Sockets;
 using System.IO.Ports;
+using CmriSubroutines.Transports;
 
 namespace CmriSubroutines
 {
@@ -403,159 +404,3 @@ namespace CmriSubroutines
     }
 }
 
-namespace CmriSubroutines
-{
-    /// <summary>
-    /// Transport abstraction to allow swapping underlying communication mechanism.
-    /// </summary>
-    public interface ITransport : IDisposable
-    {
-        void Open();
-        void Close();
-        int ReadByte();
-        int Read(byte[] buffer, int offset, int count);
-        void Write(byte[] buffer, int offset, int count);
-        void DiscardInBuffer();
-        void DiscardOutBuffer();
-        int BytesToRead { get; }
-        int BytesToWrite { get; }
-        int ReadBufferSize { get; set; }
-        int WriteBufferSize { get; set; }
-    }
-
-    /// <summary>
-    /// Serial port transport that wraps System.IO.Ports.SerialPort
-    /// </summary>
-    public class SerialTransport : ITransport
-    {
-        private readonly SerialPort _port;
-
-        public SerialTransport(int comPort, int baud100, int bufferSize)
-        {
-            if (comPort < 1)
-                throw new ArgumentOutOfRangeException(nameof(comPort));
-
-            _port = new SerialPort("COM" + comPort)
-            {
-                BaudRate = baud100 * 100,
-                Parity = Parity.None,
-                DataBits = 8,
-                StopBits = StopBits.Two,
-                ReadBufferSize = bufferSize,
-                WriteBufferSize = bufferSize
-            };
-        }
-
-        public int ReadBufferSize { get => _port.ReadBufferSize; set => _port.ReadBufferSize = value; }
-        public int WriteBufferSize { get => _port.WriteBufferSize; set => _port.WriteBufferSize = value; }
-        public int BytesToRead => _port.BytesToRead;
-        public int BytesToWrite => _port.BytesToWrite;
-
-        public void Open() => _port.Open();
-        public void Close() { if (_port.IsOpen) _port.Close(); }
-        public void Dispose() => Close();
-        public void DiscardInBuffer() => _port.DiscardInBuffer();
-        public void DiscardOutBuffer() => _port.DiscardOutBuffer();
-        public int ReadByte() => _port.ReadByte();
-        public int Read(byte[] buffer, int offset, int count) => _port.Read(buffer, offset, count);
-        public void Write(byte[] buffer, int offset, int count) => _port.Write(buffer, offset, count);
-    }
-
-    /// <summary>
-    /// TCP transport using TcpClient. Supports basic raw TCP (ser2net raw mode).
-    /// </summary>
-    public class TcpTransport : ITransport
-    {
-        private readonly TcpClient _client = new TcpClient();
-        private NetworkStream _stream;
-        private readonly string _host;
-        private readonly int _port;
-        private int _readBufferSize = 4096;
-        private int _writeBufferSize = 4096;
-
-        public TcpTransport(string host, int port)
-        {
-            _host = host ?? throw new ArgumentNullException(nameof(host));
-            _port = port;
-        }
-
-        public int ReadBufferSize { get => _readBufferSize; set => _readBufferSize = value; }
-        public int WriteBufferSize { get => _writeBufferSize; set => _writeBufferSize = value; }
-        public int BytesToRead
-        {
-            get
-            {
-                if (_stream == null || !_stream.DataAvailable) return 0;
-                return _client.Available;
-            }
-        }
-
-        public int BytesToWrite => 0; // TcpClient doesn't expose pending bytes to write easily
-
-        public void Open()
-        {
-            _client.Connect(_host, _port);
-            _stream = _client.GetStream();
-        }
-
-        public void Close()
-        {
-            try { _stream?.Close(); } catch { }
-            try { _client?.Close(); } catch { }
-        }
-
-        public void Dispose() => Close();
-
-        public void DiscardInBuffer()
-        {
-            // consume available data
-            if (_stream == null) return;
-            while (_stream.DataAvailable)
-            {
-                var buf = new byte[ReadBufferSize];
-                _stream.Read(buf, 0, buf.Length);
-            }
-        }
-
-        public void DiscardOutBuffer()
-        {
-            // nothing to do for TCP
-        }
-
-        public int ReadByte()
-        {
-            if (_stream == null) throw new InvalidOperationException("Transport not open");
-            int val = _stream.ReadByte();
-            return val;
-        }
-
-        public int Read(byte[] buffer, int offset, int count)
-        {
-            if (_stream == null) throw new InvalidOperationException("Transport not open");
-            return _stream.Read(buffer, offset, count);
-        }
-
-        public void Write(byte[] buffer, int offset, int count)
-        {
-            if (_stream == null) throw new InvalidOperationException("Transport not open");
-            _stream.Write(buffer, offset, count);
-            _stream.Flush();
-        }
-    }
-
-    /// <summary>
-    /// Simple factory to create transports based on a config object.
-    /// </summary>
-    public static class TransportFactory
-    {
-        public static ITransport CreateSerial(int comPort, int baud100, int bufferSize)
-        {
-            return new SerialTransport(comPort, baud100, bufferSize);
-        }
-
-        public static ITransport CreateTcp(string host, int port)
-        {
-            return new TcpTransport(host, port);
-        }
-    }
-}
